Issue #844 Move ValueSuppliers to internal
